2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* rtems/score/cpu.h: Remove warning in _CPU_Context_Initialize.

+++ cpukit/score/cpu/h8300/rtems/score/cpu.h
@@ -762,12 +762,12 @@ uint32_t   _CPU_ISR_Get_level( void );
                                    _isr, _entry_point, _is_fp ) \
   /* Locate Me */ \
   do { \
-    uint32_t   _stack; \
+    uintptr_t   _stack; \
     \
     if ( (_isr) ) (_the_context)->ccr = CPU_CCR_INTERRUPTS_OFF; \
     else          (_the_context)->ccr = CPU_CCR_INTERRUPTS_ON; \
     \
-    _stack = ((uint32_t)(_stack_base)) + (_size) - 4; \
+    _stack = ((uintptr_t)(_stack_base)) + (_size) - 4; \
     *((proc_ptr *)(_stack)) = (_entry_point); \
      (_the_context)->er7     = (void *) _stack; \
      (_the_context)->er6     = (void *) _stack; \
